Primus Lisp enhancements (#798)
* publishes the interrupt operation

* provides jumping observation and resolve_xxx functions

The jumping observation is made just before a jump is taken and has
two parameters: jump condition and jump destination. This is a much
more precise and useful observation than the `{enter,leave}-term`
one, as it doesn't loose the origin of values that define the
jump (i.e., we can track (taint) both the condition and the
destination).

We also added three functions to the Linker interface, mostly for
convenience. These functions resolve code names into addresses,
symbolic names, and tid.

* rectifies the primus-lisp plugin

So far the plugin loads some stuff during the link time, some after
the configuration is available, and the rest during the analysis phase
when it is invoked as a pass. It may even be invoked  twice.

Now it is rectified. Everything that doesn't require an access to the
project data structure is loaded automatically just after the
configuration phase has finished. The rest (e.g., linking with the
program) is done during the analysis.

* adds an optional fill parameter to the allocate-memory primitive

if the parameter is not set, then the allocated memory chunk is not
filled (i.e., filled with random values), if it is set, then we fill
the chunk with the specified value.

The motivation, is to make calloc more efficient.

* adds the parameter definition abstract representation

* adds parameters to the program data structure

and applies the reindexer to all runtime definitions.

* prints methods and parameters when printing a program

previously only definitions were printed

* adds defparameters form to the parser specification

* modifies defconstant grammar to fit with C. Lisp and defparameter

To be in line with Common Lisp we've decided to put docstring and
declarations after the value of defparameter and defconstant. This
indeed looks more natural.

* more efficient Lisp stack and a bug fix in signal handlers

1. Provides an efficient stack that now is equipped with a map data
structure that allows us effeciently query for locally bound
variables.

In Primus Lisp variables live in two scope:
 1. the Primus global such as CPU registers, parameters, etc.
 2. the Primus Lisp local scope that is populated by lexically scoped
    local variables such as function parameters and let bound
    variables.

Any Lisp operation on a variable involves checking whether this is a
local variable or global variable. Since we're using lexical scoping
we should be able to distinguish between them in compile time. And
that's what we will eventually do in the future. But so far it
requires to many changes to the interpreter. Previously, we were using
an assoc list, that was traversed every time we read or set a
variable. The new implementation uses a balanced tree that maps
variables to the total number of their occurences on the stack. That
makes the check to be logarithmic in the number of unique variable
names on the stack.

Few other small optimizations:

a) faster function frame costructor, that is not monadic (as it
shouldn't be), and that is not preserving the order of arguments (so
that we can append them in the reversed order, using fold left), and
that counts the total number of arguments, so that we don't call
List.length when we're poping the frame off.

b) a faster frame pushing mechanism - instead of just appending
arguments, we are prepending them in the reverse order. This is not
only faster, but since the frame is in the reverse order already, it
preserves the existing order (so no changes in semantics)

2. Fixes an nasty bug in the signal dispatcher. When a signal is
dispatched to several methods each method ignores effects from other
methods as it resets the state as it was before the dispatcher was
entered.

* keeps primitives during linking and moves init_env to loader

1. we do not want primitives, that were registered by other components
to be dropped when a new program is linked. In fact, primitives is a
complete different beast than other definitions, as they are defined
by OCaml modules (i.e., plugins) rather than by the Lisp code, so
probably they shouldn't be a part of the program data structure at
all. Though having them in the program makes things much easier.

Besides, when we drop primitives during the program linking time, we
are discarding all primitives that were registered prior to the call
to program_link. So far, just by accident, the link_program function
was called before any other components (since it was registered last -
after all other lisp components). This is so fragile, that I consider
this is a bug. Nasty one.

2. The `init_env` function in the Primus.Lisp module was adding
special variables of the form `%<id>` and `@<id>` that hold addresses
of the correspondingly named program terms. It was using the local
variables stack. Neither should it be using the local stack for
storing this variables (as they are globals by their nature), nor this
should be done during the linking time of the program. This function
is now moved to the loader (that is the component that's responsible
for setting up the environemnt).

* adds the runtime support for parameters

parameters are lazy, and are not evaluated unless are used, and not
set before.

* parametrize malloc allocator

This implementation adds the following features, that are useful
during the analysis:

- the upper limit to the maximum size of the allocated memory chunk
- the upper limit to the maximum size of the malloc heap
- malloc guards with optional coloring
- efficient calloc that doesn't
  a) take memory
  b) take time

* publishes Primus Lisp eval functions

There are two entities in Primus Lisp that could be evaluated -
functions and methods. Both are now runable directly from OCaml via
the `eval_fun` and `eval_method`.

* adds Primus Lisp autodoc generator

so far we can generate only index, but still not that bad.

* adds the documentation option that renders primus lisp documentation

* normalizes the generated documentation

- merges different definitions of the same entity
- removes quotations
- removes duplicating whitespaces

* adds documentation to various primus lisp definitions

* fixes the name initialization in the loader

an address can have multiple addresses, but a name shall have only
one, not vice verse

* drops the primus-lisp dependency