ignores return statements
this is a big change actually, and we will return to it later, to make
it more robust.

If we won't ignore them, then we will have double returns, the first
one when a function is returned based on the IR return statement (if
such is present and is well defined) and the second one, when function
exec finishes and we finally call the return destination. That's was a
disaster, and this is the fast solution. I will evolve it later