changes the Primus memory representation to preserve value identity (#806)
Historically, our implementation of the memory storage was a mapping
from bitvectors to bitvectors, thus any store operation was forgetful
as the identity of a value was discarded (where a value in Primus is a
tuple of data and unique identifier). For a long time we were
considering this as a feature, articulating that `x` and `load
a (store a x)` are two different values, even though, they are usually
structuraly equal. However, it was also observed, that in many cases
during the analysis we would like to preserve this identities, thus we
were forced to maintain our own mappings between words and value
identities in this or that way.

The proposed solution should satisfy both camps, those who believe
that `x` and `load a (store a x)` are in general different, and those
who say, that in practice they are equal as in Leibniz equality. The
memory component will now store values instead of bitvectors, so it
won't loose any information. At the same time components may define
semantics of the load/store operations as they wish, e.g., to
implement a mapping of control registers to memory, and other
non-trivial memory representations and mechanisms.

By default, however, the interpreter will ensure that `load a (store a
x)` and `x` has the same idenity if `x` is a byte. If `x` is a word of
an arbitrary length then it will be projected into several bytes with
the extract operation during the store operation, and assembled back
with the concat operation durintg the load operation.