fixes a few bugs in Primus (#818)
* skips evaluations of the mem variable in the store operation

since this variable doesn't have any meaningfull value

* fixes the cleanup procedure after term evaluation

in case of abnormal termination the cleanup procedure wasn't called.

* ignores return statements

this is a big change actually, and we will return to it later, to make
it more robust.

If we won't ignore them, then we will have double returns, the first
one when a function is returned based on the IR return statement (if
such is present and is well defined) and the second one, when function
exec finishes and we finally call the return destination. That's was a
disaster, and this is the fast solution. I will evolve it later

* adds machine-switch and machine-fork observations

also simplifies swith and fork operators, for some reason we were
holding a machine identifier in the stored continuation, probably
this is a reminiscent of the debugging stage, in any case we dont
need it anymore.

* halts machine after switch ion the greedy scheduler

it shouldn't have any observable effect, though it is better to make
it explicit.

* disables call bypassing

When we make a call we add a failsafe machine that will resume the
computation in case if the call doesn't return. We shouldn't use
this failsafe machine if everything went allright.

* handles empty blocks with branches correctly

* few cosmetic changes

* fixes the parameter order in linked lisp stubs

One of the previous commits [1] in PR #798 was claiming that we do not
need to preserve the order of the arguments, so we can provide a more
efficient frame allocation.

Apparently, we do need to preserve the order, at least of the proper
signaling the `call` and `call-return` messages. This commit reversed
the order of the arguments, since they were reversed.

[1]: https://github.com/BinaryAnalysisPlatform/bap/pull/798/commits/6d5afcd4b8dd2de1cdb03cbac12b22c633838fbf

* fixes strncpy, strncmp and adds strpbrk summaries

* parametrizes memory allocator with the zero-sentinel

that's a value that is returned when malloc is called with the zero
argument.

* implements a buffer overflow check

This check is a part of the memcheck check suite and it verifies that
functions from the string.h API are called correctly. The correctness
property is the following, if the pointer to the begining of a string
that is passed to the function, belongs to some heap region, then the
pointer to the end of this string must belong to the same heap region