fixes the compare macro in Primus Lisp
not that the macro was wrong, but it was not doing what was
expected. The old version had type `'a -> 'a -> 'a`, i.e., when
it is applied to a byte it will return a byte, and so on. This made
using this macro a little bit hard, as we need to apply casting as
usually we need just int, so our new type is `'a -> 'a -> int`.