propagates branch computations in the promiscuous mode
This is actually a major breakthrough, as this issue was preventing
taint analysis to work in promiscuous mode as it was killing all
taints on the branch conditions.

The long story - in the promiscuous mode we're flip-floping branch
conditions to reach as much code as possible. However, since we're
just assigning `1` to the branch condition we loose all taints
associated with the existing condition (or any other information that
other analyses can add).

The solution is to derive the new value from the old one. I'm using
`1 \/ c` if we need to assign `1` and `0 /\ c` if we need to assign
0.