a few fixes to primus et al (#831)
* renames eval_cond to eval-cond

All our observations, plugings, options, passes attributes, and other
textual identifiers use lisp-style dash separated notation, so
`eval_cond` looks very inconsistent.

* makes parameter evaluation visible

So far we were bypassing the interpreter when Lisp parameters were
evaluated. However, we would like to see them as first-class
computations.

In particular it helps us to track which values are used undefined.

* adds a generic taint-attached observation

this observation occurs every time a taint is attached to a value.

* fixes the compare macro in Primus Lisp

not that the macro was wrong, but it was not doing what was
expected. The old version had type `'a -> 'a -> 'a`, i.e., when
it is applied to a byte it will return a byte, and so on. This made
using this macro a little bit hard, as we need to apply casting as
usually we need just int, so our new type is `'a -> 'a -> int`.

* propagates branch computations in the promiscuous mode

This is actually a major breakthrough, as this issue was preventing
taint analysis to work in promiscuous mode as it was killing all
taints on the branch conditions.

The long story - in the promiscuous mode we're flip-floping branch
conditions to reach as much code as possible. However, since we're
just assigning `1` to the branch condition we loose all taints
associated with the existing condition (or any other information that
other analyses can add).

The solution is to derive the new value from the old one. I'm using
`1 \/ c` if we need to assign `1` and `0 /\ c` if we need to assign
0.

* changes the machine spawning algorithm in the run pass

The previous implementation was spawing machines by unhierarchical
depth-first order and was essentially creating 2^n resumptions.
The new aproach uses BFS, is linear, and, more important, is easier to
reason about.

* fixes a typo in the sign macro