fixes the kill operation in the Machine implementation
The kill operation is a hard beast, as it is very hard to suicide,
i.e., to kill the current machine. It is also hard to routinely kill
other machines in the scheduler, as usually during the rescheduling we
are in the context of a machine that we would like to kill, and
switching to another machine will take control from us.

Previous implementation of kill was just wrong, it was killing the
state without switching the continuations. In case of the self-killing
it basically was switching the state while leaving the machine in the
same continuation.

The proposed implementation, delays the machine killing if it is the
current machine, and performs it on a next fork.

Note, an easier to use (and to implement) primitive, would be [die x]
that will kill the current machine and tranfer the control to the
machine [x]. However, we can't add new primitives to the Primus
Machine interface without bumping the version. So let's keep it for
the future Primus 2.0