changes the order in which machines a run by the run plugin
When all-subroutines are selected, then we're running them in the
topological order, starting with those functions that are marked as
entry points (usually the _start function in case of an executable).

Also added an option to run each machine in isolation (i.e., in a
separate computation), might be useful for someone.