Several primus fixes (#841)
* fixes the order of the trace events

previously we were leaving a function after the first block (and all
other operations attached to the function leaving event were performed
after the entry block)

* fixes the kill operation in the Machine implementation

The kill operation is a hard beast, as it is very hard to suicide,
i.e., to kill the current machine. It is also hard to routinely kill
other machines in the scheduler, as usually during the rescheduling we
are in the context of a machine that we would like to kill, and
switching to another machine will take control from us.

Previous implementation of kill was just wrong, it was killing the
state without switching the continuations. In case of the self-killing
it basically was switching the state while leaving the machine in the
same continuation.

The proposed implementation, delays the machine killing if it is the
current machine, and performs it on a next fork.

Note, an easier to use (and to implement) primitive, would be [die x]
that will kill the current machine and tranfer the control to the
machine [x]. However, we can't add new primitives to the Primus
Machine interface without bumping the version. So let's keep it for
the future Primus 2.0

* few documentation updates (totally irrelevant to Primus)

sorry for putting it here, just don't want to create a separate PR for
such a minor stuff

* adds a couple of new functions to the posix interface

* prevents memory leaking by killing finished machines

Previously, we were waiting until the end and didn't kill any
machines.

The proposed code will kill finished machines in the greedy
scheduler. This saves tons of gigabytes of memory.

* fixes a bug in the mark-visited plugin, adds progress reports

The progress report will show the percentage of visited basic blocks,
basically the code coverage.

Also, previously marker was broken as it was marking all terms.

* removes the memory corruption check

memory corruption check was verifying that a freed pointer was
allocated. In real life it has an incredible amount of false
positives.

* fixes a bug in the primus's warn-unused analysis

also extends it a bit.

* adds machine numbers to the incident backtraces

in general it is very useful to see where the backtrace changes the
machine id, in case of the greedy scheduler and promiscuous mode those
are exactly the points where infeasibility could be introduced.

* changes the order in which machines a run by the run plugin

When all-subroutines are selected, then we're running them in the
topological order, starting with those functions that are marked as
entry points (usually the _start function in case of an executable).

Also added an option to run each machine in isolation (i.e., in a
separate computation), might be useful for someone.