rectifies reconstructor, symtab and brancher
This PR rectifies reconstructor, symtab and brancher with a respect to
performance, without adding new behaviour or breaking of existed one.

Rewrote it and reduced a number of iterations over nodes/edges of cfg

There was a bit inefficient implementation of `add_symbol` function,
so every addition of a symbol led to a filter of the whole table,
although there are enough info to reduce such calls.

Just a small fix that check if an instruction has jumps at all before
subsequent call of `fold_consts` that could be heavy for some of
instructions