improves dead-code-elimination
This PR intends to speed up the dead-code-elimination plugin.
At first, it adds cache, so every next launch will use a result
of a first one.

At second, and this need to be discussed IRL, this reduces a number
of SSA form calculation, that subsequently speed up the whole plugin