makes param abstract and embeds digest into it
To track changes in informational content of parameters we use a
digest function, which injects the domain of a parameter into the
space of hashes.

Note, the hash function is cryptographically (or stochastically)
injective, in a sense that althtough it is not strictly injective,
and it is possible to have digest(p) = digest(q) /\ q <> p, however
probability of obtaining such pair (q,p) is neglible.

The approach allows us to represent our computation as a merkle tree,
that opens an opportunity for efficient caching.