implements deadcode-elimination pass optimization  (#857)
* improves dead-code-elimination

This PR intends to speed up the dead-code-elimination plugin.
At first, it adds cache, so every next launch will use a result
of a first one.

At second, and this need to be discussed IRL, this reduces a number
of SSA form calculation, that subsequently speed up the whole plugin

* bug fixed

* bug fixed again

* updated after review

Sorry, indented the buffer, so a bit more effort could be required
for review.

changes:
1) consts propagation and dead code elimination separated one from
   another. Previously, the number of times we called `propagate_consts`
   was the same as number of times we did dead code elimination.
   Theoreticly speaking, these changes could kill more unused
   code, since we propagate consts until reaching of a fixpoint,
   and only then eliminate dead code.
2) caching takes in account results of consts propagation too
3) refactoring

* refactoring and bug fixing

This PR brings back interleaving of constant propagation
and dead code elimination. Also optimizes constant propagation
in part of computing set of virtual variables for propagation:
it's done only once now.

* const propagation in ssa form

We run const propagation in ssa form, so we don't need to track once defined
variable, since they are obviously defined only once in ssa.
Added a cmdline flag for propagating consts in real variables too,
i.e. in registers.

* adds a level of optimization

Now, set of variables that are affectd by analysis could
be defined through the comamnd line

* refactored and reduced memory usage

This PR reduces a memory usage by applying analysis to
each subroutine separatly, so there are not multiply copies
of the same subroutine in memory, like it was before.

* added attribute for updated terms

* refactoring

* fixed a bug in sub's digest

Sub name was't taken in account, so digest could be the same
for different (but similar) subroutines

* renamed dead-code-elimination to optimization

* updated testsuite

* bug fixed

* feqw updates after review

* edited doc

* issue a warning on --no-dead-code-elimination