adds constant tracking analysis
The analysis tracks static constants in a binary, i.e., values
are hardcoded in a program.

For the purposes of the analysis a value is a static constant if
it was initialized from a constant value or computed from a static
constant value.

The constant tracker analysis can be called from Primus Lisp using
the `all-static-constants` and `points-to-static-data` functions.
See `--primus-lisp-documentation` for the documentation on this
functions.

Along with the constant tracker this PR adds some sample analysis.

For example, check-hardcoded-values reports the
`hardcoded-socket-address` incident every time a hardcoded socket
address is used in one of the BSD Socket API function.

Another analysis is implemented in the `check-null-pointers` module
that reports a `null-pointer-dereference` incident every time a static
NULL pointer is dereferenced.