silenced warnings from ours deprecations
This PR silences warnings from bap code during compilation of
bap itself. Also, it turned out that few modules in `bap.mli`
were deprecated in a way that doesn't emit warnings at all:

Just for the record, the next doesn't deprecate `module A`,
because according to the documentation for attributes of level 3:
> They are not attached to any specific node in the syntax tree

```
module A = struct
  [@@@deprecated "[since ...]"]
  ...
end
```

And the next works as expected, although in this case we have
do the deprecation in the end of module, which makes it less
readable. But anyway, looks like we don't have any other choice.

```
module A = struct
  ...
end [@@deprecated "[since ...]"]
```