ensures static linking with llvm >= 4.0
it turned out that after `llvm-4.0` a default linking mode is `shared`,
and it's not what we want. e.g. debian package will be broken, as `bap`
will be depend from `LLVM.so`.

our previous request for libraries now returns an extremly short list:
```
ubuntu@ubuntu-xenial:~$ llvm-config-4.0 --libs
-lLLVM-4.0
```
which is just shared library.

So what we can do, is to use a new flag for llvm-config:
```
ubuntu@ubuntu-xenial:~$ llvm-config-4.0 --link-static --libs
-lLLVMLTO -lLLVMPasses -lLLVMObjCARCOpts -lLLVMMIRParser ...

```
Here it is!