allows for construction of values from raw semantics.
In other words, adds `Value.create : 'a sort -> semantics -> 'a value`
and `Eff.create : 'a kind -> semantics -> 'a eff` functions, so that
the semantics of a program might be denoted in terms of abstract
semantics of its elements.

Given a concrete example, we can denote the semantics of a floating
point value in terms of the semantics of its constituents which are
bitvectors, e.g.,

```ocaml
type gfloat = {
  sign : semantics;
  expn : semantics;
  koef : semantics;
  flag : float_state;
}
```
Now, since all terms are sorted, we can always safely reconstruct the
value from the semantics:

```ocaml
let significand : ('e,'s) float value t -> 's bitv value t =
  fun x -> x >>= fun v ->
  let s = Floats.sigs (Value.sort v) in
  let {koef} = Value.get gfloat v in
  Knowledge.return (Value.create s koef)
```

Note, the type annotation here is for the readers reference only and
is not needed, moreover the type system will infer it correctly.

Also, it is advised to use helper functions, such as `>>->` to get the
sort and payload in one call, see the existing lifters for the
example.