provides a fresh Label generator in the Knowledge module
also weakens the requirements for the contents declaration,
adds a public flag and makes desc optional. The name is required
to be unique only if content is made public. So that the private
content providers don't need to bother about descriptions and unique
names. Still work in progress, so expect the interface to change.