adds the floating points grammar and parser
and also adds `cast_sint` as a primitive to the floating point
signature, as well as fixes the `sqrt` primitive, by renaming
it to `fsqrt` (for the consistency with the rest of the floating point
primitives) and making it a unary operation (well... because it is).

In the grammar we now have a new syntactic class - rmode. Technically,
the shouldn't be but we can allow ourselves this kind of flexibility.