adds the is_virtual predicate to the variables theory
a variable is virtual if it represents a temporary storage for the
result of computation. A special syntax `$N` is used to represent
them, and the only possible way to obtain them is to use the
Var.Generator module.

Notice, this commit also adds a validation routine that prevents users
from constructing values with that syntax manually, as well as
disables some other invalid identifiers, e.g., those that contain non
letters. Maybe it is a little bit too strict, we can relax it later if neeed.