abstracts floating point representation and revaps the float theory
as representing floating point explicitly leads to an explosion of our
expressions we've decided to review our approach to floating point
representation.

First of all, floating point now is viewed as a special kind of a
bitvector, not as an abstract arithmetic type.

We drop arithmetic constructors and destructors and propose instead
a constructor parametrized by abstract floating point format, and
only one destructor that returns a bitvector that represents the
floating point in the given format.

The floating point sort is not longer parametrized by the exponent and
significand size but with the representation size and a format. Not
only this allows us to express floating point arithmetics without
ascribing the exponents and signficands, but it also makes arithmetic
operations between floating points with different representations not
well sorted.

The new representation enables any floating points, such as IEEE, MIL
STD, IBM, or TI, or whatever you want. However, since IEEE is so
popular we decided to put it directly into the Core Theory.

The Theory itself doesn't depend on IEEE, but the provided
parser (which is a covenience tool for uplifting BIL-like languages
into the Core Theory) depends on IEEE formats (and it is the only
format that it understands). We took  the liberty, however, and put
the extended double format (aka x87, aka binary80) into the set of
IEEE formats. Althogh, officially it is not a part of the standard,
our abstractions allow for it.