makes it sound
Unfortunately, the original proposal was unsound, as the assumption
that only the last definition should be preserved is implied by the
assumption that only the last definition is used for interprocedural
communication, which is totally wrong.

The liveness property is propagated from each call (and indirect jump)
to all members of the protected set. We're using the SSA form for
def-use analysis which is inherently intraprocedural, so to preserve
the liveness of the variables with the interprocedural scope we use
the protected set as an overapproximation.  A tighter approximation
would be to reinstantiate their liveness after each call or indirect
jump. However, this would require tampering with the ssa algorithm,
that we don't want to do right now.

The updated solution reinstantiates liveness of protected variables
at the end of each block, even the block is not a call or even jump.
The liveness is killed by the first use of protection, so it is
propagated only to the last (if any) definition.