pushes semantics from the term level down to leaf terms
not sure what the semantics of a term should mean, so I've pushed
the semantics to individual leaf terms. It amounted to lots of work
at the end of the day, since we need to provide backward compatibility
as well as support lots of operations on terms.

Published, so far, only the new semantics interface for jmps and defs,
will later publish the rest.